+++ src/lib/wasm_lib_wrapper.ts
